bring "Quick Install" to the top of the README
This pulls the Quick Install section, which most users will be looking
for, to the top of the README.  It also adds a sentence about ensuring
you have python 3 and pip.

The pre-requisites section was confusing to me because it starts by
saying I might be okay with python 2 or 3 for the hook, but then says
I need python 3 anyway for the registration script.  Therefore I
didn't bother mentioning python 2 at all in the new sentence.

The original pre-requesites section is still present, but I think
maybe it could be removed.  Python 2 is EOL anyway.